+++ apps/buyer-backend-spring/src/main/java/ccm/buyer/service/AuctionService.java
@@ -1,7 +1,9 @@
 package ccm.buyer.service;
 
+import java.math.BigDecimal;
+
 import ccm.buyer.entity.Bid;
 
 public interface AuctionService {
-    Bid placeBid(Long buyerId, Long auctionId, Double amount);
+    Bid placeBid(Long buyerId, Long auctionId, BigDecimal amount);
 }
