+++ apps/buyer-backend-spring/src/main/java/ccm/buyer/service/BuyerService.java
@@ -1,7 +1,9 @@
 package ccm.buyer.service;
 
+import java.math.BigDecimal;
+
 import ccm.buyer.entity.Transaction;
 
 public interface BuyerService {
-    Transaction directBuy(Long buyerId, Long listingId, Integer qty);
+    Transaction directBuy(Long buyerId, Long listingId, BigDecimal qty);
 }
